+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+
+## 1.7.1 - 08-10-2021
+
+* fix bug resume download on Android
+* upgrade Android dependencies (WorkerManager 2.7.0-rc01)
+
+## 1.7.0 - 07-09-2021
+
+* support Android 11 (ScopedStorage, PackageVisibility)
+
+## 1.6.1 - 29-05-2021
+
+* fix bug on Android: `getContentLength`
+* upgrade Android dependencies (WorkerManager 2.5.0)
+
+## 1.6.0 - 12-04-2021
+
+* Stable version of nullsafety
+
+## 1.6.0-nullsafety.0 - 28.02.2021
+
+* Fix bugs
+* Migrate to nullsafety
+
+## 1.5.2 - 25.10.2020
+
+* Android: fix bug notification stuck in processing
+
+## 1.5.1 - 27.09.2020
+
+* iOS: fix bug missing update download progress
+
 ## 1.5.0 - 09.08.2020
 
 * Update `pubspec` to new format